Renamed some of the files to clarify axi_lite and axi_stream modules.
