+++ veriutils/__init__.py
@@ -2,4 +2,3 @@ from .cosimulation import *
 from .hdl_blocks import *
 from .utils import *
 from .axi import *
-from .axi_lite import *
